+++ src/main/java/com/example/aereopuerto/service/FavoritoService.java
@@ -63,6 +63,10 @@ public class FavoritoService {
     public void removeFavorito(Integer personaId, Integer favoritoId) {
         Favorito favorito = favoritoRepository.findById(favoritoId)
                 .orElseThrow(() -> new FavoritoInvalidoException("Favorito no encontrado con id: " + favoritoId));
+        if (!favorito.getPersona().getId().equals(personaId)) {
+            throw new FavoritoInvalidoException("El favorito no pertenece a esta persona.");
+        }
+
         favoritoRepository.delete(favorito);
     }
 
@@ -75,7 +79,7 @@ public class FavoritoService {
         return dto;
     }
 
-    @CacheEvict(value = "favoritos", key = "#result.personaId") // <-- ¡Ojo a este truco para la caché!
+    @CacheEvict(value = "favoritos", key = "#result.personaId")
     public FavoritoDTO addFavoritoPorToken(String email, Integer vueloId) {
 
         User user = userRepository.findByEmail(email)
